fix: Use default text style in DSUserAvatar to fix overflow
