+++ sample/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:blip_ds/blip_ds.dart';
 import 'package:flutter/material.dart';
 import 'package:get/route_manager.dart';
-import 'package:sample/widgets/showcase/sample_user_avatar.showcas.dart';
+import 'package:sample/widgets/showcase/sample_user_avatar.showcase.dart';
 
 import 'widgets/showcase/sample_bottom_sheet.showcase.dart';
 import 'widgets/showcase/sample_button.showcase.dart';
